packages/builder/my solution/requirements.txt
streamlit
google-generativeai
openai
SpeechRecognition